chore: refactor top songs implementation to reflect design guidelines
