+++ lib/private/SPOTIFY_ENDPOINTS.ts
@@ -1,2 +1,6 @@
+// USER INFO
 export const TOKEN = 'https://accounts.spotify.com/api/token';
 export const ME = 'https://api.spotify.com/v1/me';
+
+// SONGS
+export const TOP_TRACKS = 'https://api.spotify.com/v1/me/top/tracks';
